# docs/swagger.json
{"swagger":"2.0","info":{"title":"GinChat","contact":{},"version":"1.0"},"paths":{"/contact/addfriend":{"post":{"tags":["聊天界面"],"summary":"添加好友","parameters":[{"type":"string","description":"用户ID","name":"userId","in":"query"},{"type":"string","description":"好友ID","name":"targetId","in":"query"}],"responses":{"200":{"description":"code, massage, data\"}","schema":{"type":"string"}}}}},"/contact/createCommunity":{"post":{"tags":["群聊界面"],"summary":"新建群聊","parameters":[{"type":"string","description":"用户ID","name":"ownerId","in":"query"},{"type":"string","description":"群名称","name":"name","in":"query"},{"type":"string","description":"群头像","name":"icon","in":"query"},{"type":"string","description":"群描述","name":"desc","in":"query"}],"responses":{"200":{"description":"code, massage, data\"}","schema":{"type":"string"}}}}},"/contact/loadcommunity":{"post":{"tags":["群聊界面"],"summary":"加载群列表","parameters":[{"type":"string","description":"用户ID","name":"ownerId","in":"query"}],"responses":{"200":{"description":"code, massage, data\"}","schema":{"type":"string"}}}}},"/index":{"get":{"tags":["首页"],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}},"/searchFriends":{"post":{"tags":["聊天界面"],"summary":"查询好友列表","parameters":[{"type":"string","description":"用户ID","name":"userId","in":"query"}],"responses":{"200":{"description":"code, massage, data\"}","schema":{"type":"string"}}}}},"/toChat":{"get":{"tags":["IM用户界面"],"responses":{}}},"/toRegister":{"get":{"tags":["用户注册"],"responses":{}}},"/user/createUser":{"get":{"tags":["用户模块"],"summary":"新增用户","parameters":[{"type":"string","description":"用户名","name":"name","in":"query"},{"type":"string","description":"密码","name":"password","in":"query"},{"type":"string","description":"确认密码","name":"repasswd","in":"query"}],"responses":{"200":{"description":"code\", \"massage\", \"data\"}","schema":{"type":"string"}}}}},"/user/deleteUser":{"get":{"tags":["用户模块"],"summary":"注销用户","parameters":[{"type":"string","description":"id","name":"id","in":"query"}],"responses":{"200":{"description":"code\", \"massage\", \"data\"}","schema":{"type":"string"}}}}},"/user/findUserByNameAndPwd":{"post":{"tags":["用户模块"],"summary":"用户登录校验","parameters":[{"type":"string","description":"用户名","name":"name","in":"query"},{"type":"string","description":"密码","name":"password","in":"query"}],"responses":{"200":{"description":"code\", \"massage\", \"data\"}","schema":{"type":"string"}}}}},"/user/getUserList":{"get":{"tags":["用户模块"],"summary":"查看所有用户","responses":{"200":{"description":"code\", \"massage\", \"data\"}","schema":{"type":"string"}}}}},"/user/updateUser":{"post":{"tags":["用户模块"],"summary":"修改用户信息","parameters":[{"type":"string","description":"id","name":"id","in":"formData"},{"type":"string","description":"name","name":"name","in":"formData"},{"type":"string","description":"password","name":"password","in":"formData"},{"type":"string","description":"phone","name":"phone","in":"formData"},{"type":"string","description":"email","name":"email","in":"formData"}],"responses":{"200":{"description":"code\", \"massage\", \"data\"}","schema":{"type":"string"}}}}}}}